% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R
\name{cjpowr_amce}
\alias{cjpowr_amce}
\title{Calculate Power or Minimum Required Sample Size for AMCEs in Conjoint Experiments}
\usage{
cjpowr_amce(
  amce,
  alpha = 0.05,
  power = NULL,
  levels = 2,
  treat.prob = 0.5,
  n = NULL,
  sims = 1e+05
)
}
\arguments{
\item{amce}{of a user-specified size.}

\item{alpha}{the significance level, 0.05 by default.}

\item{power}{power of the test (1 minus Type II error probability), default is NULL.}

\item{levels}{the number of levels of the treatment.}

\item{treat.prob}{default is 0.5.}

\item{n}{effective sample size, default is NULL.}

\item{sims}{number of simulation runs to compute the expected type M error ("exaggeration ratio"), default 100k. If NULL, E(type M) is not computed.}
}
\description{
Based on Schuessler/Freitag (2020), this function calculates the minimum effective
sample size (MES) required to reject a null hypothesis of no effect with probablity (power) if the alternative is true for a given average marginal component effect (AMCE) and a given alpha level.
The function can be manipulated to give power for a given effective sample size. Type S (incorrect sign)and the expected type M errors (exaggeration ratio) are also computed (Lu/Qui/Deng 2019).
The effective sample size is the number of profiles times the number of tasks. The outcome is assumed to be binary (forced-choice).
}
\details{
This function assumes a binary outcome variable and the AMCE recast in a regression setting using a two-sided t-test.
Using conservative assumptions, the MES can simply be divided by
the number of profiles times the number of tasks to optimize the design vis-a-vis
resource constraints. The number of attributes does not affect power.
The function is deliberately simple and its output can easily be passed.

Passing instead the effective sample size to the function gives power.
Type S (incorrect sign)and the expected type M errors (exaggeration ratio) are also computed (Lu/Qui/Deng 2019). If your study is
not completely underpowered, you do not need to worry about type S error. However, only high powered
studies result in low E(type M erorrs).

To calculate power for conditional AMCEs or interaction effects,
users are referred to \link[cjpowR]{cjpowr_amcie}.
}
\section{Literature}{


1. Schuessler, J. and M. Freitag (2020). Power Analysis for Conjoint Experiments.
   Working Paper.

2. Lu, J., Y. Qiu, and A. Deng (2019). A note on Type S/M errors in hypothesis testing.
   British Journal of Mathematical and Statistical Psychology 72(1), 1–17.
}

\examples{
# This gives the minimum required effective sample size (type S, E(type M)):
df = cjpowr_amce(amce = 0.05, power = 0.8, levels = 5)

# For example, for a conjoint with 2 profiles and 4 tasks, n becomes:

df$n/(2*4)

#This gives the power (type S, E(type M)):
cjpowr_amce(amce = 0.05, n = 7829.258, levels = 5)

#Generating an interactive plot for type M error:

d <- expand.grid(amce = c(0.01, 0.02, 0.03, 0.05), n = seq(from = 100, to = 50000, length.out = 1000))

# Purrr Style:
library(purrr)
set.seed(123)
df <- pmap_df(d, function(amce, n) cjpowr_amce(amce = amce, n = n, sims = 1000, levels = 5, alpha = 0.05, treat.prob = 0.5))

# Base R:
set.seed(123)
cjpowr_amce_vec <- Vectorize(cjpowr_amce)
df2 <- t(cjpowr_amce_vec(amce = d$amce, n = d$n, sims = 1000, levels = 5, alpha = 0.05, treat.prob = 0.5))

df2 <- data.frame(df2)
df2[] <- lapply(df2, unlist)

library(plotly)

plot_ly(df, x = ~n, y = ~exp_typeM, type = 'scatter', mode = 'lines', linetype = ~amce) \%>\%
 layout(
   xaxis = list(title = "Effective Sample Size",
                zeroline = F,
                hoverformat = '.0f'),
   yaxis = list(title = "Exaggeration Ratio",
                range = c(0,10),
                zeroline = F,
                hoverformat = '.2f'),
   legend=list(title=list(text='<b> AMCE </b>')),
   hovermode = "x unified"
 )
}
\keyword{AMCE}
\keyword{analysis,}
\keyword{conjoint,}
\keyword{power}
