% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcie.R
\name{cjpowr_amcie}
\alias{cjpowr_amcie}
\title{Calculate Power or Minimum Required Sample Size for AMCIEs/Conditional AMCEs in Conjoint Experiments}
\usage{
cjpowr_amcie(
  delta0 = 0.5,
  delta1 = 0,
  delta3,
  alpha = 0.05,
  power = NULL,
  n = NULL,
  levels1 = 2,
  levels2 = 2,
  treat.prob = list(p00 = 0.25, p10 = 0.25, p01 = 0.25, p11 = 0.25),
  sims = 1e+05
)
}
\arguments{
\item{delta0}{default is 0.5.}

\item{delta1}{default is 0.}

\item{delta3}{amcie or conditional amce of user specified size.}

\item{alpha}{the significance level, 0.05 by default.}

\item{power}{power of the test (1 minus Type II error probability), default is NULL.}

\item{n}{effective sample size, default is NULL.}

\item{levels1}{the number of levels of the first treatment. Default is 2.}

\item{levels2}{the number of levels of the second treatment/subgroup variable. Default is 2.}

\item{treat.prob}{List of treatment probabilities. If the interest is in causal interaction and uniform randomization is employed, then p00=p10=p01=p11= 0.25. Thus, the default is list(0.25,0.25,0.25,0.25)}

\item{sims}{number of simulation runs to compute the expected type M error ("exaggeration ratio"), default 100k. If NULL, E(type M) is not computed.}
}
\description{
Based on Schuessler/Freitag (2020), this function calculates the minimum effective
sample size (MES) required to reject a null hypothesis of no effect with probablity (power) if the alternative is true for a given average marginal component interaction effect (AMCIE) and a given alpha level.
Function can be manipulated to give power for a given effective sample size. Type S (incorrect sign)and the expected type M errors (exaggeration ratio) are also computed (Lu/Qui/Deng 2019).
The effective sample size is the number of profiles times the number of tasks. The outcome is assumed to be binary (forced-choice).
}
\details{
This function assumes a binary outcome variable and the AMCIE recast in a regression setting using a two-sided t-test.
Using conservative assumptions, the MES can simply be divided by
the number of profiles times the number of taks to optimize the design vis-a-vis
resource constraints. The number of attributes does not affect power.

To calculate power or minimum sample size for differences in conditional AMCEs, given the known or estimated marginal distribution of the pre-treatment co-variate,
the argument "treat.prob" can easily be manipulated. For instance, interacting a uniformly randomized treatment with a
binary co-variate with e.g. P_covariate(W_ijk = 0) = 4/5 yields the following list of probabilities: list(0.1,0.1,0.4,0.4). Details on the formula can be found in Schuessler/Freitag 2020.
Shifting more probability mass on p01 and p11 (as in this example) results in marginally higher MES / lowe power (see the technical Appendix of the paper).

Passing the effective sample size to the function gives power.
Type S (incorrect sign)and the expected type M errors (exaggeration ratio) are also computed (Lu/Qui/Deng 2019). If your study is
not completely underpowered, you do not need to worry about type S error. However, only high powered
studies result in low E(type M errors).

Power for AMCEs can be calculated with \link[cjpowR]{cjpowr_amce}.
}
\section{Literature}{


1. Schuessler, J. and M. Freitag (2020). Power Analysis for Conjoint Experiments.
   Working Paper.

2. Lu, J., Y. Qiu, and A. Deng (2019). A note on Type S/M errors in hypothesis testing.
   British Journal of Mathematical and Statistical Psychology 72(1), 1–17.
}

\examples{

cjpowr_amcie(delta3 = 0.05, levels1=3, levels2=3, power=0.8)

#Generating an interactive plot for type M error:

cjpowr_amcie_vec <- Vectorize(cjpowr_amcie)
d <- expand.grid(delta3 = c(0.01, 0.02, 0.03, 0.05), n = seq(from = 100, to = 50000, length.out = 1000))
df <- t(cjpowr_amcie_vec(delta3 = d$delta3, n = d$n, sims = 10000, levels1 = 3, levels2=4, alpha = 0.05))
df <- data.frame(df)
df[] <- lapply(df, unlist)
plot_ly(df, x = ~n, y = ~exp_typeM, type = 'scatter', mode = 'lines', linetype = ~delta3) \%>\%
 layout(
   xaxis = list(title = "Effective Sample Size",
                zeroline = F,
                hoverformat = '.0f'),
   yaxis = list(title = "Exaggeration Ratio",
                range = c(0,10),
                zeroline = F,
                hoverformat = '.2f'),
   legend=list(title=list(text='<b> AMCIE </b>')),
   hovermode = "x unified"
 )
}
\keyword{AMCE}
\keyword{analysis,}
\keyword{conjoint,}
\keyword{power}
