% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amce.R
\name{cjpowr_plotly}
\alias{cjpowr_plotly}
\title{Interactive Power Curves for AMCEs}
\usage{
cjpowr_plotly(
  y,
  x,
  by,
  power = NULL,
  levels = NULL,
  alpha = NULL,
  amce = NULL,
  n = NULL,
  delta0 = 0.5
)
}
\arguments{
\item{y}{axis of the interactive plot.}

\item{x}{axis of the interactive plot.}

\item{by}{third dimension of the 2d plot.}

\item{power}{see details.}

\item{levels}{see details.}

\item{alpha}{see details.}

\item{amce}{see details.}

\item{n}{see details.}

\item{delta0}{see details.}
}
\description{
This is a convenience function to quickly visualize basic power curves for AMCEs with plotly.
The function will be extended to provide power curves for AMCIEs or conditional AMCEs in the future.
But see \link[cjpowR]{cjpowr_amcie} for plotting examples. Returns a \link[plotly]{plot_ly} object that can be easily manipulated.
}
\details{
In this early version, you can specify y only as "power" or "n" and x as "amce" or "n". Whatever has been set as x, has to be passed as a sequence of values
specified by seq(). For instance, if you want a range of AMCEs on the x axis, argument amce has to be = seq(from = 0.02, to = 0.1, length.out = 1000).
Whatever has been set as y does not have to be set as argument and cannot be set to x.
To add a third dimension to the 2d plot, specify by with one of the remaining parameters. Whatever has been
set to by, has to be passed as a list of values.
For all remaining parameters, set single values. At the moment, delta0 cannot be set as "by" or "x".
}
\examples{
cjpowr_plotly(y = "n", x = "amce", by = "power",
amce = seq(from = 0.02, to = 0.1, length.out = 1000),
power = c(0.7,0.8,0.9),
alpha = 0.05,
levels = 5,
delta0=0.5)

cjpowr_plotly(y = "power", x = "n", by = "amce",
amce = c(0.02, 0.03, 0.04),
n = seq(from = 100, to = 30000, by = 5),
alpha = 0.05,
levels = 5,
delta0=0.5)
}
\keyword{AMCE}
\keyword{analysis,}
\keyword{conjoint,}
\keyword{power}
